fix(api/users): ページビューのデータの取得範囲が狭すぎる問題を修正
＋リファクタリング
